requirements.txt
# DV3 Voice Companion - Dependencies

# Gemini Live API
google-genai>=1.0.0

# Audio
sounddevice>=0.4.6
numpy>=1.24.0

# Wake Word (use ONNX backend — tflite-runtime lacks Python 3.12+ wheels)
openwakeword>=0.6.0
onnxruntime>=1.16.0

# Visualizer
pygame>=2.5.0
Pillow>=10.0.0

# Spotify
spotipy>=2.23.0

# Config & Environment
python-dotenv>=1.0.0
PyYAML>=6.0

# Editor (uses same Pillow + Pygame as visualizer)
# No additional deps

# WebSocket server (web visualizer)
aiohttp>=3.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
